refactor: AdminPage, MainPage 불필요 함수 제거
